+++ src/locales/numberLocales/index.ts
