+++ src/locales/timeLocales/index.ts
