+++ src/utils/locale.ts
@@ -10,8 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
-import { numberLocale } from 'locales/numberLocales/numberLocales';
-import { timeLocale } from 'locales/timeLocales/timeLocales';
+import { numberLocale } from 'locales/numberLocales';
+import { timeLocale } from 'locales/timeLocales';
 import { ChartProps } from 'types';
 import { NumberLocaleCode, TimeLocaleCode } from 'types/locales';
 import { Locale, NumberLocale, TimeLocale } from 'vega';
